% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPdokTiles.R
\name{addPdokTiles}
\alias{addPdokTiles}
\title{Leaflet function to add pdok tiles}
\usage{
addPdokTiles(map, type = c("brt", "aerial", "pastel", "gray"), ...)
}
\arguments{
\item{map}{leaflet object}

\item{type}{one of the PDOK tiles: currently: "brt", "aerial", "pastel" or "gray"}

\item{...}{Passed to \code{\link{leaflet::addTiles}}.}
}
\description{
Leaflet function to add pdok tiles
}
