% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_geocode.R
\name{nl_geocode}
\alias{nl_geocode}
\title{geocode function}
\usage{
nl_geocode(location, output = c("latlon", "latlona", "rd", "rda", "all"),
  messaging = FALSE, type = NULL, ..., verbose = !messaging)
}
\description{
Which is more user friendly than the barebones webservices,
uses the same function signature as ggmap geocode.
}
\details{
The return type can be specified.
}
