% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_lookup.R
\name{nl_lookup}
\alias{nl_lookup}
\title{Look up a geo object}
\usage{
nl_lookup(id, ..., output = c("list", "raw"), verbose = FALSE)
}
\arguments{
\item{id}{of object found in nl_suggest or nl_free}

\item{...}{extra parameters are passed to the lookup service of pdok.}

\item{output}{What type of output should be returned}

\item{verbose}{should the function print message while retrieving the data?}
}
\description{
Retrieve detailed properties of a geo object found with \code{\link{nl_suggest}} or
\code{\link{nl_free}}.
}
\examples{
\dontrun{

obj <- nl_lookup("weg-f633e85f07eda4e68a00fb13f9d128f5")
names(obj)
obj$weergavenaam

}
}
