% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest.R
\name{nl_suggest}
\alias{nl_suggest}
\title{Get a list of suggestion for geolocations in NL}
\usage{
nl_suggest(q, ..., type = NULL, verbose = TRUE)
}
\arguments{
\item{q}{search terms that should be geolocated}

\item{...}{parameters passed to geolocation service}

\item{type}{restrict type of geolocation to}
}
\description{
Get a list of suggestion for geolocations in NL
}
