% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nlgeocoder-package}
\alias{nlgeocoder}
\alias{nlgeocoder-package}
\title{nlgeocoder package}
\description{
The nlgeocoder package provides a R wrapper around the pdok webservice.
}
\details{
it allows to geocode data using the following functions:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uRos2018/nlgeocoder}
  \item Report bugs at \url{https://github.com/uRos2018/nlgeocoder/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com}

Authors:
\itemize{
  \item Willy Tadema
  \item Egge-Jan Pollé
}

Other contributors:
\itemize{
  \item Juris Breidaks [contributor]
}

}
