% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Dummy function only used for describing common api parameters}
\usage{
query(rows = NULL, start = NULL, wt)
}
\arguments{
\item{rows}{number of rows

#' @seealso}
}
\description{
Dummy function only used for describing common api parameters
}
\keyword{internal}
