% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_geocode.R
\name{nl_geocode}
\alias{nl_geocode}
\alias{nl_geocode_rd}
\alias{nl_geocode_df}
\title{geocode function}
\usage{
nl_geocode(location, output = c("wgs84", "rd", "data.frame"),
  messaging = FALSE, type = "adres", ..., verbose = messaging)

nl_geocode_rd(location, messaging = FALSE, type = "adres", ...,
  verbose = messaging)

nl_geocode_df(location, messaging = FALSE, type = "adres", ...,
  verbose = messaging)
}
\arguments{
\item{location}{string with location to be found}

\item{output}{Should the output be a \code{data.frame} or
\code{\link[sf]{sf}} object in wgs84 or Rijksdriehoekstelsel format?}

\item{messaging}{Print the urls fired to the webserver (consistent with `ggmap::geocode`)}

\item{type}{retrict the type of object that is returned from the service.}

\item{...}{will be passed to \code{\link{nl_free}}.}

\item{verbose}{identical to \code{messaging} (consistent with other nlgeoder
functions)}
}
\description{
This function is more user friendly than the barebones webservices,
and uses the same function signature as ggmap::geocode.
}
\details{
type can be one of the following:"provincie", "gemeente", "woonplaats",
weg;postcode;adres;perceel;hectometerpaal;wijk;buurt;waterschapsgrens;appartementsrecht.
The return type can be specified and can be of type "sf" or "data.frame".
}
