% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_lookup.R
\name{nl_lookup}
\alias{nl_lookup}
\title{Look up a geo object}
\usage{
nl_lookup(id, ..., raw = FALSE, verbose = FALSE)
}
\arguments{
\item{id}{of object found in nl_suggest or nl_free}
}
\description{
Retrieve detailed properties of a geo object found with suggest or free.
}
