% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_suggest.R
\name{nl_suggest}
\alias{nl_suggest}
\title{Get a list of suggestions for geolocations in NL}
\usage{
nl_suggest(q, ..., rows = NULL, type = NULL, verbose = FALSE,
  fl = NULL, sort = NULL, qf = NULL, bq = NULL)
}
\arguments{
\item{q}{search terms that should be geolocated}

\item{...}{parameters passed to geolocation service}

\item{rows}{maximum number of rows to be returned. Default is 10.}

\item{type}{restrict type of geolocation to}

\item{verbose}{Should the functions print messages on what it is retrieving.}

\item{fl}{the columns that should be returned (aka select on columms of result)}

\item{sort}{how the data should be sorted}

\item{qf}{the fields that should be queried}

\item{bq}{the boosting of the query.}
}
\description{
Get a list of suggestions for geolocations in NL
}
