% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nlgeocoder-package}
\alias{nlgeocoder}
\alias{nlgeocoder-package}
\title{Geo coder package}
\description{
R interface to the location server API of PDOK.
}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com}

Authors:
\itemize{
  \item Willy Tadema
  \item Egge-Jan Pollé
}

}
