% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nlgeocoder-package}
\alias{nlgeocoder}
\alias{nlgeocoder-package}
\title{nlgeocoder package}
\description{
The nlgeocoder package provides a R wrapper around the pdok webservice.
}
\details{
it allows to geocode data using the following functions:

\itemize{
  \item \code{\link{nl_geocode}}: geocode a vector of addresses.
  \item \code{\link{nl_free}}: find possible matches for one address.
  \item \code{\link{nl_suggest}}: find possible suggestions for one address.
  \item \code{\link{nl_lookup}}: lookup information of object found in
  \code{nl_suggest}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uRos2018/nlgeocoder}
  \item Report bugs at \url{https://github.com/uRos2018/nlgeocoder/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (0000-0002-6580-4718)

Authors:
\itemize{
  \item Willy Tadema
  \item Egge-Jan Pollé
}

Other contributors:
\itemize{
  \item Juris Breidaks [contributor]
}

}
