% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{nlgeocoder-package}
\alias{nlgeocoder}
\alias{nlgeocoder-package}
\title{nlgeocoder package}
\description{
The nlgeocoder package provides a R wrapper around the pdok webservice.
}
\details{
it allows to geocode data using the following functions:
}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com}

Authors:
\itemize{
  \item Willy Tadema
  \item Egge-Jan Pollé
}

}
