% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Dummy function only used for describing common api parameters}
\usage{
query(rows = NULL, start = NULL, fq = NULL, lat = NULL,
  lon = NULL, indent = NULL, wt = "json")
}
\arguments{
\item{rows}{maximum number of rows that will be returned, if not specified it will be 10.}

\item{start}{Starting index is van het eerste resultaat dat teruggegeven wordt. Dit is
zero-based. In combinatie met de rows-parameter kunnen deze services gepagineerd worden bevraagd. De default-waarde is "0"}

\item{fq}{Hiermee kan een filter query worden opgegeven, bijv. fq=bron:BAG. Met fq=* kan de default filter query worden opgeheven. De default filter is type:(gemeente OR woonplaats OR weg OR postcode OR adres)}

\item{lat}{Hiermee kan een coördinaat (in lat/lon, ofwel WGS84) worden opgegeven. Met behulp van deze parameters worden de gevonden zoekresultaten gesorteerd op afstand van het meegegeven punt. Wanneer de locatie van de gebruiker bekend is, kan op deze manier effectiever worden gezocht.
Het meegeven van een coördinaat is met name nuttig voor de suggest- en vrije geocoder-services. Hier worden meestal meerdere resultaten teruggegeven. Als decimaal scheidingsteken moet een punt worden opgegeven i.p.v. een komma.}

\item{lon}{see lat}

\item{indent}{Hiermee kan worden opgegeven of de teruggegeven JSON ingesprongen (geïndenteerd) moet worden. Geldige waarden zijn "true" en "false". De default-waarde is "true".}

\item{wt}{Hiermee wordt opgegeven wat het outputformaat is van de bevraging. Geldige waarden zijn "json" en "xml". De default-waarde is "json".}
}
\description{
Dummy function only used for describing common api parameters
}
\keyword{internal}
