% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Dummy function only used for describing common api parameters}
\usage{
query(rows = NULL, start = NULL, fq = NULL, lat = NULL,
  lon = NULL, indent = NULL, wt = "json")
}
\arguments{
\item{rows}{maximum number of rows to be returned. Default is 10.}

\item{start}{start index.Default is 0.}

\item{fq}{apply a filter to the query, e.g. fq=bron:BAG. The default filter is 'type:(municipality OR town OR neiborhood OR postcode OR adress)'. Use fq=* to remove the default filter.}

\item{lat}{enter a decimal degree (latitude, in WGS84). Together with lon this will define a point to search from. Search results will be ordered by distance from this point. For example: 'lat=52.09&lon=5.12'}

\item{lon}{enter a decimal degree (longitude, in WGS84). Together with lat this will define  point to search from. Search results will be ordered by distance from this point. For example: 'lat=52.09&lon=5.12'}

\item{indent}{return JSON in pretty print (true) or not (false). Default is true.}

\item{wt}{output format. Possible values are json and xml. Default is json.}
}
\description{
Dummy function only used for describing common api parameters
}
\keyword{internal}
